Extract client package
