+++ common/persistence/elasticsearch/client/aws.go
@@ -22,45 +22,58 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package elasticsearch
+package client
 
 import (
 	"fmt"
 	"net/http"
+	"os"
+	"strings"
 
-	"go.temporal.io/server/common/log"
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/credentials"
+	"github.com/aws/aws-sdk-go/aws/session"
+	elasticaws "github.com/olivere/elastic/aws/v4"
+
 	"go.temporal.io/server/common/service/config"
 )
 
-func NewClient(config *config.Elasticsearch, httpClient *http.Client, logger log.Logger) (Client, error) {
-	switch config.Version {
-	case "v6", "":
-		return newClientV6(config, httpClient, logger)
-	case "v7":
-		return newClientV7(config, httpClient, logger)
-	default:
-		return nil, fmt.Errorf("not supported ElasticSearch version: %v", config.Version)
+func NewAwsHttpClient(config config.ESAWSRequestSigningConfig) (*http.Client, error) {
+	if !config.Enabled {
+		return nil, nil
 	}
-}
 
-func NewCLIClient(url string, version string) (CLIClient, error) {
-	switch version {
-	case "v6":
-		return newSimpleClientV6(url)
-	case "v7", "":
-		return newSimpleClientV7(url)
-	default:
-		return nil, fmt.Errorf("not supported ElasticSearch version: %v", version)
+	if config.Region == "" {
+		config.Region = os.Getenv("AWS_REGION")
+		if config.Region == "" {
+			return nil, fmt.Errorf("unable to resolve AWS region for obtaining AWS Elastic signing credentials")
+		}
 	}
-}
 
-func NewIntegrationTestsClient(url string, version string) (IntegrationTestsClient, error) {
-	switch version {
-	case "v6":
-		return newSimpleClientV6(url)
-	case "v7":
-		return newSimpleClientV7(url)
+	var awsCredentials *credentials.Credentials
+
+	switch strings.ToLower(config.CredentialProvider) {
+	case "static":
+		awsCredentials = credentials.NewStaticCredentials(
+			config.Static.AccessKeyID,
+			config.Static.SecretAccessKey,
+			config.Static.Token,
+		)
+	case "environment":
+		awsCredentials = credentials.NewEnvCredentials()
+	case "aws-sdk-default":
+		awsSession, err := session.NewSession(&aws.Config{
+			Region: &config.Region,
+		})
+
+		if err != nil {
+			return nil, err
+		}
+
+		awsCredentials = awsSession.Config.Credentials
 	default:
-		return nil, fmt.Errorf("not supported ElasticSearch version: %v", version)
+		return nil, fmt.Errorf("unknown AWS credential provider specified: %+v. Accepted options are 'static', 'environment' or 'session'", config.CredentialProvider)
 	}
+
+	return elasticaws.NewV4SigningClient(awsCredentials, config.Region), nil
 }
