+++ common/persistence/elasticsearch/client/bulk_processor_v6.go
@@ -22,7 +22,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package elasticsearch
+package client
 
 import (
 	elastic6 "github.com/olivere/elastic"
