+++ common/persistence/elasticsearch/client/bulk_service_v7.go
@@ -22,46 +22,51 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-//go:generate mockgen -copyright_file ../../../LICENSE -package $GOPACKAGE -source $GOFILE -destination bulk_processor_mock.go
-
-package elasticsearch
+package client
 
 import (
-	"time"
+	"context"
 
 	"github.com/olivere/elastic/v7"
 )
 
-type BulkableRequestType uint8
-
-const (
-	BulkableRequestTypeIndex BulkableRequestType = iota
-	BulkableRequestTypeDelete
-)
-
 type (
-	BulkProcessor interface {
-		Stop() error
-		Add(request *BulkableRequest)
-	}
-
-	// BulkProcessorParameters holds all required and optional parameters for executing bulk service
-	BulkProcessorParameters struct {
-		Name          string
-		NumOfWorkers  int
-		BulkActions   int
-		BulkSize      int
-		FlushInterval time.Duration
-		Backoff       elastic.Backoff
-		BeforeFunc    elastic.BulkBeforeFunc
-		AfterFunc     elastic.BulkAfterFunc
-	}
-
-	BulkableRequest struct {
-		RequestType BulkableRequestType
-		Index       string
-		ID          string
-		Version     int64
-		Doc         map[string]interface{}
+	bulkServiceV7 struct {
+		esBulkService *elastic.BulkService
 	}
 )
+
+func newBulkServiceV7(esBulkService *elastic.BulkService) *bulkServiceV7 {
+	return &bulkServiceV7{
+		esBulkService: esBulkService,
+	}
+}
+
+func (b *bulkServiceV7) Do(ctx context.Context) error {
+	_, err := b.esBulkService.Do(ctx)
+	return err
+}
+
+func (b *bulkServiceV7) NumberOfActions() int {
+	return b.esBulkService.NumberOfActions()
+}
+
+func (b *bulkServiceV7) Add(request *BulkableRequest) {
+	switch request.RequestType {
+	case BulkableRequestTypeIndex:
+		bulkDeleteRequest := elastic.NewBulkIndexRequest().
+			Index(request.Index).
+			Id(request.ID).
+			VersionType(versionTypeExternal).
+			Version(request.Version).
+			Doc(request.Doc)
+		b.esBulkService.Add(bulkDeleteRequest)
+	case BulkableRequestTypeDelete:
+		bulkDeleteRequest := elastic.NewBulkDeleteRequest().
+			Index(request.Index).
+			Id(request.ID).
+			VersionType(versionTypeExternal).
+			Version(request.Version)
+		b.esBulkService.Add(bulkDeleteRequest)
+	}
+}
