+++ common/persistence/elasticsearch/client/logger.go
@@ -1,0 +1,57 @@
+// The MIT License
+//
+// Copyright (c) 2020 Temporal Technologies Inc.  All rights reserved.
+//
+// Copyright (c) 2020 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package client
+
+import (
+	"fmt"
+
+	"go.temporal.io/server/common/log"
+)
+
+type (
+	errorLogger struct {
+		log.Logger
+	}
+
+	infoLogger struct {
+		log.Logger
+	}
+)
+
+func newErrorLogger(logger log.Logger) *errorLogger {
+	return &errorLogger{logger}
+}
+
+func (l *errorLogger) Printf(format string, v ...interface{}) {
+	l.Error(fmt.Sprintf(format, v...))
+}
+
+func newInfoLogger(logger log.Logger) *infoLogger {
+	return &infoLogger{logger}
+}
+
+func (l *infoLogger) Printf(format string, v ...interface{}) {
+	l.Info(fmt.Sprintf(format, v...))
+}
