+++ common/persistence/elasticsearch/visibility_store_write_test.go
@@ -33,6 +33,7 @@ import (
 
 	"go.temporal.io/server/common/payload"
 	"go.temporal.io/server/common/persistence"
+	"go.temporal.io/server/common/persistence/elasticsearch/client"
 	"go.temporal.io/server/common/searchattribute"
 )
 
@@ -60,7 +61,7 @@ func (s *ESVisibilitySuite) TestRecordWorkflowExecutionStarted() {
 	}
 
 	s.mockProcessor.EXPECT().Add(gomock.Any(), gomock.Any(), gomock.Any()).
-		Do(func(bulkRequest *BulkableRequest, visibilityTaskKey string, ackCh chan<- bool) {
+		Do(func(bulkRequest *client.BulkableRequest, visibilityTaskKey string, ackCh chan<- bool) {
 			s.NotNil(ackCh)
 			s.Equal(1, cap(ackCh))
 			s.Equal(0, len(ackCh))
@@ -86,7 +87,7 @@ func (s *ESVisibilitySuite) TestRecordWorkflowExecutionStarted() {
 			// %q because request has JSON encoded string.
 			s.EqualValues(request.SearchAttributes.GetIndexedFields()[searchattribute.CustomStringField].Data, fmt.Sprintf("%q", searchAttributes[searchattribute.CustomStringField]))
 
-			s.Equal(BulkableRequestTypeIndex, bulkRequest.RequestType)
+			s.Equal(client.BulkableRequestTypeIndex, bulkRequest.RequestType)
 			s.EqualValues(request.TaskID, bulkRequest.Version)
 			s.Equal("wid~rid", bulkRequest.ID)
 			s.Equal("test-index", bulkRequest.Index)
@@ -105,7 +106,7 @@ func (s *ESVisibilitySuite) TestRecordWorkflowExecutionStarted_EmptyRequest() {
 	}
 
 	s.mockProcessor.EXPECT().Add(gomock.Any(), gomock.Any(), gomock.Any()).
-		Do(func(bulkRequest *BulkableRequest, visibilityTaskKey string, ackCh chan<- bool) {
+		Do(func(bulkRequest *client.BulkableRequest, visibilityTaskKey string, ackCh chan<- bool) {
 			s.NotNil(ackCh)
 			s.Equal(1, cap(ackCh))
 			s.Equal(0, len(ackCh))
@@ -120,7 +121,7 @@ func (s *ESVisibilitySuite) TestRecordWorkflowExecutionStarted_EmptyRequest() {
 			_, ok = body[searchattribute.Encoding]
 			s.False(ok)
 
-			s.Equal(BulkableRequestTypeIndex, bulkRequest.RequestType)
+			s.Equal(client.BulkableRequestTypeIndex, bulkRequest.RequestType)
 			s.EqualValues(request.TaskID, bulkRequest.Version)
 			s.Equal("~", bulkRequest.ID)
 			s.Equal("test-index", bulkRequest.Index)
@@ -156,7 +157,7 @@ func (s *ESVisibilitySuite) TestRecordWorkflowExecutionClosed() {
 	}
 
 	s.mockProcessor.EXPECT().Add(gomock.Any(), gomock.Any(), gomock.Any()).
-		Do(func(bulkRequest *BulkableRequest, visibilityTaskKey string, ackCh chan<- bool) {
+		Do(func(bulkRequest *client.BulkableRequest, visibilityTaskKey string, ackCh chan<- bool) {
 			s.NotNil(ackCh)
 			s.Equal(1, cap(ackCh))
 			s.Equal(0, len(ackCh))
@@ -178,7 +179,7 @@ func (s *ESVisibilitySuite) TestRecordWorkflowExecutionClosed() {
 			s.EqualValues(request.Status, body[searchattribute.ExecutionStatus])
 			s.EqualValues(request.HistoryLength, body[searchattribute.HistoryLength])
 
-			s.Equal(BulkableRequestTypeIndex, bulkRequest.RequestType)
+			s.Equal(client.BulkableRequestTypeIndex, bulkRequest.RequestType)
 			s.EqualValues(request.TaskID, bulkRequest.Version)
 			s.Equal("wid~rid", bulkRequest.ID)
 			s.Equal("test-index", bulkRequest.Index)
@@ -197,7 +198,7 @@ func (s *ESVisibilitySuite) TestRecordWorkflowExecutionClosed_EmptyRequest() {
 	}
 
 	s.mockProcessor.EXPECT().Add(gomock.Any(), gomock.Any(), gomock.Any()).
-		Do(func(bulkRequest *BulkableRequest, visibilityTaskKey string, ackCh chan<- bool) {
+		Do(func(bulkRequest *client.BulkableRequest, visibilityTaskKey string, ackCh chan<- bool) {
 			s.NotNil(ackCh)
 			s.Equal(1, cap(ackCh))
 			s.Equal(0, len(ackCh))
@@ -212,7 +213,7 @@ func (s *ESVisibilitySuite) TestRecordWorkflowExecutionClosed_EmptyRequest() {
 			_, ok = body[searchattribute.Encoding]
 			s.False(ok)
 
-			s.Equal(BulkableRequestTypeIndex, bulkRequest.RequestType)
+			s.Equal(client.BulkableRequestTypeIndex, bulkRequest.RequestType)
 			s.EqualValues(request.TaskID, bulkRequest.Version)
 			s.Equal("~", bulkRequest.ID)
 			s.Equal("test-index", bulkRequest.Index)
@@ -232,7 +233,7 @@ func (s *ESVisibilitySuite) TestDeleteExecution() {
 	}
 
 	s.mockProcessor.EXPECT().Add(gomock.Any(), gomock.Any(), gomock.Any()).
-		Do(func(bulkRequest *BulkableRequest, visibilityTaskKey string, ackCh chan<- bool) {
+		Do(func(bulkRequest *client.BulkableRequest, visibilityTaskKey string, ackCh chan<- bool) {
 			s.NotNil(ackCh)
 			s.Equal(1, cap(ackCh))
 			s.Equal(0, len(ackCh))
@@ -240,7 +241,7 @@ func (s *ESVisibilitySuite) TestDeleteExecution() {
 
 			s.Equal("wid~rid", visibilityTaskKey)
 
-			s.Equal(BulkableRequestTypeDelete, bulkRequest.RequestType)
+			s.Equal(client.BulkableRequestTypeDelete, bulkRequest.RequestType)
 			s.EqualValues(request.TaskID, bulkRequest.Version)
 			s.Equal("wid~rid", bulkRequest.ID)
 			s.Equal("test-index", bulkRequest.Index)
@@ -255,7 +256,7 @@ func (s *ESVisibilitySuite) TestDeleteExecution_EmptyRequest() {
 	request := &persistence.VisibilityDeleteWorkflowExecutionRequest{}
 
 	s.mockProcessor.EXPECT().Add(gomock.Any(), gomock.Any(), gomock.Any()).
-		Do(func(bulkRequest *BulkableRequest, visibilityTaskKey string, ackCh chan<- bool) {
+		Do(func(bulkRequest *client.BulkableRequest, visibilityTaskKey string, ackCh chan<- bool) {
 			s.NotNil(ackCh)
 			s.Equal(1, cap(ackCh))
 			s.Equal(0, len(ackCh))
@@ -263,7 +264,7 @@ func (s *ESVisibilitySuite) TestDeleteExecution_EmptyRequest() {
 
 			s.Equal("~", visibilityTaskKey)
 
-			s.Equal(BulkableRequestTypeDelete, bulkRequest.RequestType)
+			s.Equal(client.BulkableRequestTypeDelete, bulkRequest.RequestType)
 			s.Equal("~", bulkRequest.ID)
 			s.Equal("test-index", bulkRequest.Index)
 		})
